Use an arguments file when fetching proto file info
When protowrap is invoked with a large number of proto files as
arguments, its call to protoc when building file descriptors can have
enough arguments to exceed system limits on argument size. This commit
fixes the issue by passing proto filenames in a file and using protoc's
"@file" feature for reading argument from the file.